Refactor parent verification and child addition logic
Moved parent verification and child addition logic from SpeciesDescription and ClassDescription to TypeDescription for better code reuse and consistency. Removed redundant hierarchy checks and streamlined the validation of parent relationships. Minor updates to error reporting and debugging initialization.
